fix: refactor downloadFileUsingRequesterPays sample into its own file
